Makefile: specify the build image, the tty interactive (#13)
Add variable BUILD_ENV, default to local. If the variable is not local,
then the TTY_IT_OPT variable is set to empty and the docker container to
build the agent is started without the allocation of a TTY for
interactive mode. This can be handy for a DIND environment.

Add variable BUILD_IMAGE, default to "profiling-agent". If specified
otherwise, this new image is used to start the docker container to build
the agent. This can be good for testing with different images and
toolchains.