INSTANA_README: fix tag creation, add patching/release info
Fix the description of how to create a tag after the synch: incorporate
the procedure of how to retrieve the last upstream commit and indicate
where exactly you need to write it in the tag message.

Add patching and release info.
The patching mechanism was completely missing from the README, now it is
explained in detail.
The Release procedure is explained and detailed with an Example, based
on the latest Release.
Details of where to find the pipelines in IBM Cloud are given.